+++ packages/@rescript/runtime/lib/js/Stdlib_Int.js
@@ -4,7 +4,7 @@ let Stdlib_Array = require("./Stdlib_Array.js");
 
 function fromString(x, radix) {
   let maybeInt = radix !== undefined ? parseInt(x, radix) : parseInt(x);
-  if (isNaN(maybeInt) || maybeInt > 2147483647 || maybeInt < -2147483648) {
+  if (Number.isNaN(maybeInt) || maybeInt > 2147483647 || maybeInt < -2147483648) {
     return;
   } else {
     return maybeInt | 0;